refactor(go): only include ResolveEnvironmentBaseURL for multi-URL environments
