+++ generators/go/internal/generator/generator.go
@@ -577,6 +577,9 @@ func (g *Generator) generate(ir *fernir.IntermediateRepresentation, mode Mode) (
 		files = append(files, newFileParamFile(g.coordinator, rootPackageName, generatedNames))
 		files = append(files, newHttpCoreFile(g.coordinator))
 		files = append(files, newHttpInternalFile(g.coordinator))
+		if isMultipleBaseUrlsEnvironment(ir.Environments) {
+			files = append(files, newEnvironmentInternalFile(g.coordinator))
+		}
 		files = append(files, newPointerFile(g.coordinator, rootPackageName, generatedNames))
 		files = append(files, newQueryFile(g.coordinator))
 		files = append(files, newQueryTestFile(g.coordinator))
@@ -1256,6 +1259,14 @@ func newHttpInternalFile(coordinator *coordinator.Client) *File {
 		coordinator,
 		"internal/http.go",
 		[]byte(httpInternalFile),
+	)
+}
+
+func newEnvironmentInternalFile(coordinator *coordinator.Client) *File {
+	return NewFile(
+		coordinator,
+		"internal/environment.go",
+		[]byte(environmentInternalFile),
 	)
 }
